Update quests.yaml
Did my best to summarize the very long and complicated questline of Millicent. I haven't gotten to the end of the game yet myself, so the end of it is the best I could surmise from the wiki - please feel free to correct it or raise an issue if you spot mistakes.