Update sorceries.yaml + Reorder Items pages
Redid the numbering for all the new list items to include renumbering the first spell in each broken up batch. All the untouched spells have the same numbers and everything that's been touched has a chronological number that wasn't in its original list.
I also included the alphabetization of the Items pages. It includes a commented out line for the page I'm working on now.