Added images to many pages, moved many images into folders
Added images to every page in the Upgrades & Unlocks and Collectables tabs. Some pages have images that apply to whole sections, not individual items; in these cases, I linked the section title to the image (commented out if there was already another link on the section title) to address later.