Like 20% of the armor images
The icons/armor folder is being sorted into "done" for everything that's added to the page so far, "player_alterable", and "unusable". So far all are incomplete. In the end the "done" folder will get deleted and its contents dumped back into the armor folder, it's just there to keep track.